+++ src/rdkafka.h
@@ -167,7 +167,7 @@ typedef SSIZE_T ssize_t;
  * @remark This value should only be used during compile time,
  *         for runtime checks of version use rd_kafka_version()
  */
-#define RD_KAFKA_VERSION 0x020800ff
+#define RD_KAFKA_VERSION 0x020900ff
 
 /**
  * @brief Returns the librdkafka version as integer.
@@ -288,7 +288,7 @@ typedef enum {
         RD_KAFKA_RESP_ERR__BAD_MSG = -199,
         /** Bad/unknown compression */
         RD_KAFKA_RESP_ERR__BAD_COMPRESSION = -198,
-        /** Broker is going away */
+        /** Broker is going away, together with client instance */
         RD_KAFKA_RESP_ERR__DESTROY = -197,
         /** Generic failure */
         RD_KAFKA_RESP_ERR__FAIL = -196,
@@ -412,6 +412,8 @@ typedef enum {
         /** A different record in the batch was invalid
          *  and this message failed persisting. */
         RD_KAFKA_RESP_ERR__INVALID_DIFFERENT_RECORD = -138,
+        /** Broker is going away but client isn't terminating */
+        RD_KAFKA_RESP_ERR__DESTROY_BROKER = -137,
 
         /** End internal error codes */
         RD_KAFKA_RESP_ERR__END = -100,
@@ -654,6 +656,9 @@ typedef enum {
         /** Client sent a push telemetry request larger than the maximum size
          *  the broker will accept. */
         RD_KAFKA_RESP_ERR_TELEMETRY_TOO_LARGE = 118,
+        /** Client metadata is stale,
+         *  client should rebootstrap to obtain new metadata. */
+        RD_KAFKA_RESP_ERR_REBOOTSTRAP_REQUIRED = 129,
         RD_KAFKA_RESP_ERR_END_ALL,
 } rd_kafka_resp_err_t;
 
@@ -5371,6 +5376,18 @@ void rd_kafka_group_list_destroy(const struct rd_kafka_group_list *grplist);
 RD_EXPORT
 int rd_kafka_brokers_add(rd_kafka_t *rk, const char *brokerlist);
 
+/**
+ * @brief Retrieve and return the learned broker ids.
+ *
+ * @param rk Instance to use.
+ * @param cntp Will be updated to the number of brokers returned.
+ *
+ * @returns a malloc:ed list of int32_t broker ids.
+ *
+ * @remark The returned pointer must be freed.
+ */
+RD_EXPORT
+int32_t *rd_kafka_brokers_learned_ids(rd_kafka_t *rk, size_t *cntp);
 
 
 /**