+++ src/rdkafka_mock.c
@@ -49,6 +49,8 @@ static void rd_kafka_mock_cluster_destroy0(rd_kafka_mock_cluster_t *mcluster);
 static rd_kafka_mock_request_t *
 rd_kafka_mock_request_new(int32_t id, int16_t api_key, int64_t timestamp_us);
 static void rd_kafka_mock_request_free(void *element);
+static void rd_kafka_mock_coord_remove(rd_kafka_mock_cluster_t *mcluster,
+                                       int32_t broker_id);
 
 static rd_kafka_mock_broker_t *
 rd_kafka_mock_broker_find(const rd_kafka_mock_cluster_t *mcluster,
@@ -437,16 +439,22 @@ rd_kafka_mock_partition_assign_replicas(rd_kafka_mock_partition_t *mpart,
         int replica_cnt = RD_MIN(replication_factor, mcluster->broker_cnt);
         rd_kafka_mock_broker_t *mrkb;
         int i = 0;
-        int first_replica =
-            (mpart->id * replication_factor) % mcluster->broker_cnt;
+        int first_replica;
         int skipped = 0;
 
         if (mpart->replicas)
                 rd_free(mpart->replicas);
 
-        mpart->replicas    = rd_calloc(replica_cnt, sizeof(*mpart->replicas));
+        mpart->replicas = replica_cnt
+                              ? rd_calloc(replica_cnt, sizeof(*mpart->replicas))
+                              : NULL;
         mpart->replica_cnt = replica_cnt;
+        if (replica_cnt == 0) {
+                rd_kafka_mock_partition_set_leader0(mpart, NULL);
+                return;
+        }
 
+        first_replica = (mpart->id * replication_factor) % mcluster->broker_cnt;
 
         /* Use a predictable, determininistic order on a per-topic basis.
          *
@@ -938,6 +946,23 @@ static void rd_kafka_mock_cluster_io_add(rd_kafka_mock_cluster_t *mcluster,
         mcluster->fd_cnt++;
 }
 
+/**
+ * @brief Reassign partition replicas to broker, after deleting or
+ *        adding a new one.
+ */
+static void
+rd_kafka_mock_cluster_reassign_partitions(rd_kafka_mock_cluster_t *mcluster) {
+        rd_kafka_mock_topic_t *mtopic;
+        TAILQ_FOREACH(mtopic, &mcluster->topics, link) {
+                int i;
+                for (i = 0; i < mtopic->partition_cnt; i++) {
+                        rd_kafka_mock_partition_t *mpart =
+                            &mtopic->partitions[i];
+                        rd_kafka_mock_partition_assign_replicas(
+                            mpart, mpart->replica_cnt);
+                }
+        }
+}
 
 static void rd_kafka_mock_connection_close(rd_kafka_mock_connection_t *mconn,
                                            const char *reason) {
@@ -1584,10 +1609,36 @@ static void rd_kafka_mock_broker_destroy(rd_kafka_mock_broker_t *mrkb) {
         if (mrkb->rack)
                 rd_free(mrkb->rack);
 
+        rd_kafka_mock_coord_remove(mrkb->cluster, mrkb->id);
+
         TAILQ_REMOVE(&mrkb->cluster->brokers, mrkb, link);
         mrkb->cluster->broker_cnt--;
 
         rd_free(mrkb);
+}
+
+
+rd_kafka_resp_err_t
+rd_kafka_mock_broker_decommission(rd_kafka_mock_cluster_t *mcluster,
+                                  int32_t broker_id) {
+        rd_kafka_op_t *rko = rd_kafka_op_new(RD_KAFKA_OP_MOCK);
+
+        rko->rko_u.mock.broker_id = broker_id;
+        rko->rko_u.mock.cmd       = RD_KAFKA_MOCK_CMD_BROKER_DECOMMISSION;
+
+        return rd_kafka_op_err_destroy(
+            rd_kafka_op_req(mcluster->ops, rko, RD_POLL_INFINITE));
+}
+
+rd_kafka_resp_err_t rd_kafka_mock_broker_add(rd_kafka_mock_cluster_t *mcluster,
+                                             int32_t broker_id) {
+        rd_kafka_op_t *rko = rd_kafka_op_new(RD_KAFKA_OP_MOCK);
+
+        rko->rko_u.mock.broker_id = broker_id;
+        rko->rko_u.mock.cmd       = RD_KAFKA_MOCK_CMD_BROKER_ADD;
+
+        return rd_kafka_op_err_destroy(
+            rd_kafka_op_req(mcluster->ops, rko, RD_POLL_INFINITE));
 }
 
 
@@ -1684,16 +1735,28 @@ static int rd_kafka_mock_broker_new_listener(rd_kafka_mock_cluster_t *mcluster,
 
 
 static rd_kafka_mock_broker_t *
-rd_kafka_mock_broker_new(rd_kafka_mock_cluster_t *mcluster, int32_t broker_id) {
+rd_kafka_mock_broker_new(rd_kafka_mock_cluster_t *mcluster,
+                         int32_t broker_id,
+                         rd_kafka_resp_err_t *err) {
         rd_kafka_mock_broker_t *mrkb;
         rd_socket_t listen_s;
         struct sockaddr_in sin = {
             .sin_family = AF_INET,
             .sin_addr   = {.s_addr = htonl(INADDR_LOOPBACK)}};
 
-        listen_s = rd_kafka_mock_broker_new_listener(mcluster, &sin);
-        if (listen_s == -1)
+        if (rd_kafka_mock_broker_find(mcluster, broker_id)) {
+                if (err)
+                        *err = RD_KAFKA_RESP_ERR__INVALID_ARG;
+                /* A broker with this id already exists. */
                 return NULL;
+        }
+
+        listen_s = rd_kafka_mock_broker_new_listener(mcluster, &sin);
+        if (listen_s == -1) {
+                if (err)
+                        *err = RD_KAFKA_RESP_ERR__TRANSPORT;
+                return NULL;
+        }
 
         /*
          * Create mock broker object
@@ -1718,6 +1781,8 @@ rd_kafka_mock_broker_new(rd_kafka_mock_cluster_t *mcluster, int32_t broker_id) {
 
         if (rd_kafka_mock_broker_start_listener(mrkb) == -1) {
                 rd_kafka_mock_broker_destroy(mrkb);
+                if (err)
+                        *err = RD_KAFKA_RESP_ERR__TRANSPORT;
                 return NULL;
         }
 
@@ -1824,6 +1889,20 @@ rd_kafka_mock_coord_set(rd_kafka_mock_cluster_t *mcluster,
         TAILQ_INSERT_TAIL(&mcluster->coords, mcoord, link);
 
         return mcoord;
+}
+
+/**
+ * @brief Remove coordinator by broker id.
+ */
+void rd_kafka_mock_coord_remove(rd_kafka_mock_cluster_t *mcluster,
+                                int32_t broker_id) {
+        rd_kafka_mock_coord_t *mcoord, *tmp;
+
+        TAILQ_FOREACH_SAFE(mcoord, &mcluster->coords, link, tmp) {
+                if (mcoord->broker_id == broker_id) {
+                        rd_kafka_mock_coord_destroy(mcluster, mcoord);
+                }
+        }
 }
 
 
@@ -2237,6 +2316,31 @@ rd_kafka_mock_broker_set_rack(rd_kafka_mock_cluster_t *mcluster,
             rd_kafka_op_req(mcluster->ops, rko, RD_POLL_INFINITE));
 }
 
+void rd_kafka_mock_broker_set_host_port(rd_kafka_mock_cluster_t *cluster,
+                                        int32_t broker_id,
+                                        const char *host,
+                                        int port) {
+        rd_kafka_mock_broker_t *mrkb;
+
+        mtx_lock(&cluster->lock);
+        TAILQ_FOREACH(mrkb, &cluster->brokers, link) {
+                if (mrkb->id == broker_id) {
+                        rd_kafka_dbg(
+                            cluster->rk, MOCK, "MOCK",
+                            "Broker %" PRId32
+                            ": Setting advertised listener from %s:%d to %s:%d",
+                            broker_id, mrkb->advertised_listener, mrkb->port,
+                            host, port);
+                        rd_snprintf(mrkb->advertised_listener,
+                                    sizeof(mrkb->advertised_listener), "%s",
+                                    host);
+                        mrkb->port = port;
+                        break;
+                }
+        }
+        mtx_unlock(&cluster->lock);
+}
+
 rd_kafka_resp_err_t
 rd_kafka_mock_coordinator_set(rd_kafka_mock_cluster_t *mcluster,
                               const char *key_type,
@@ -2358,6 +2462,11 @@ rd_kafka_mock_broker_cmd(rd_kafka_mock_cluster_t *mcluster,
                         mrkb->rack = NULL;
                 break;
 
+        case RD_KAFKA_MOCK_CMD_BROKER_DECOMMISSION:
+                rd_kafka_mock_broker_destroy(mrkb);
+                rd_kafka_mock_cluster_reassign_partitions(mcluster);
+                break;
+
         default:
                 RD_BUG("Unhandled mock cmd %d", rko->rko_u.mock.cmd);
                 break;
@@ -2412,6 +2521,7 @@ rd_kafka_mock_cluster_cmd(rd_kafka_mock_cluster_t *mcluster,
         rd_kafka_mock_partition_t *mpart;
         rd_kafka_mock_broker_t *mrkb;
         size_t i;
+        rd_kafka_resp_err_t err;
 
         switch (rko->rko_u.mock.cmd) {
         case RD_KAFKA_MOCK_CMD_TOPIC_CREATE:
@@ -2526,8 +2636,16 @@ rd_kafka_mock_cluster_cmd(rd_kafka_mock_cluster_t *mcluster,
         case RD_KAFKA_MOCK_CMD_BROKER_SET_UPDOWN:
         case RD_KAFKA_MOCK_CMD_BROKER_SET_RTT:
         case RD_KAFKA_MOCK_CMD_BROKER_SET_RACK:
+        case RD_KAFKA_MOCK_CMD_BROKER_DECOMMISSION:
                 return rd_kafka_mock_brokers_cmd(mcluster, rko);
 
+        case RD_KAFKA_MOCK_CMD_BROKER_ADD:
+                if (!rd_kafka_mock_broker_new(mcluster,
+                                              rko->rko_u.mock.broker_id, &err))
+                        return err;
+
+                rd_kafka_mock_cluster_reassign_partitions(mcluster);
+                break;
         case RD_KAFKA_MOCK_CMD_COORD_SET:
                 if (!rd_kafka_mock_coord_set(mcluster, rko->rko_u.mock.name,
                                              rko->rko_u.mock.str,
@@ -2646,15 +2764,15 @@ static void rd_kafka_mock_cluster_destroy0(rd_kafka_mock_cluster_t *mcluster) {
 
         rd_list_destroy(&mcluster->request_list);
 
+        dummy_rkb_thread = mcluster->dummy_rkb->rkb_thread;
+
         /*
-         * Destroy dummy broker
+         * Destroy dummy broker.
+         * WARNING: This is last time we can read
+         * from dummy_rkb in this thread!
          */
         rd_kafka_q_enq(mcluster->dummy_rkb->rkb_ops,
                        rd_kafka_op_new(RD_KAFKA_OP_TERMINATE));
-
-        dummy_rkb_thread = mcluster->dummy_rkb->rkb_thread;
-
-        rd_kafka_broker_destroy(mcluster->dummy_rkb);
 
         if (thrd_join(dummy_rkb_thread, &ret) != thrd_success)
                 rd_assert(!*"failed to join mock dummy broker thread");
@@ -2728,7 +2846,7 @@ rd_kafka_mock_cluster_t *rd_kafka_mock_cluster_new(rd_kafka_t *rk,
         TAILQ_INIT(&mcluster->brokers);
 
         for (i = 1; i <= broker_cnt; i++) {
-                if (!(mrkb = rd_kafka_mock_broker_new(mcluster, i))) {
+                if (!(mrkb = rd_kafka_mock_broker_new(mcluster, i, NULL))) {
                         rd_kafka_mock_cluster_destroy(mcluster);
                         return NULL;
                 }