+++ src/rdkafka_mock.h
@@ -335,6 +335,22 @@ rd_kafka_mock_broker_set_down(rd_kafka_mock_cluster_t *mcluster,
                               int32_t broker_id);
 
 /**
+ * @brief Sets a new \p host and \p port for a given broker identified by
+ *        \p broker_id.
+ *
+ * @param mcluster Mock cluster instance.
+ * @param broker_id The id of the broker to modify.
+ * @param host The new hostname.
+ * @param port The new port.
+ */
+RD_EXPORT void
+rd_kafka_mock_broker_set_host_port(rd_kafka_mock_cluster_t *mcluster,
+                                   int32_t broker_id,
+                                   const char *host,
+                                   int port);
+
+
+/**
  * @brief Makes the broker accept connections again.
  *        This does NOT trigger leader change.
  *
@@ -368,6 +384,33 @@ rd_kafka_mock_broker_set_rack(rd_kafka_mock_cluster_t *mcluster,
                               int32_t broker_id,
                               const char *rack);
 
+
+
+/**
+ * @brief Remove and delete a mock broker from a cluster.
+ *        All partitions assigned to that broker will be
+ *        reassigned to other brokers.
+ *
+ * @param cluster The mock cluster containing the broker
+ * @param broker_id The broker to delete
+ * @returns 0 on success or -1 on error
+ */
+RD_EXPORT int
+rd_kafka_mock_broker_decommission(rd_kafka_mock_cluster_t *cluster,
+                                  int32_t broker_id);
+
+/**
+ * @brief Add a new broker to the cluster.
+ *        Cluster partition will be reassigned to use the new broker
+ *        as well.
+ *
+ * @param mcluster The mock cluster
+ * @param broker_id The id of the broker to add
+ *
+ * @returns Error value or 0 if no error occurred
+ */
+RD_EXPORT rd_kafka_resp_err_t
+rd_kafka_mock_broker_add(rd_kafka_mock_cluster_t *mcluster, int32_t broker_id);
 
 
 /**